# интерфейсы/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Кейс 02: регистрация, оплата и доступ",
    "description": "Исполняемая демонстрация проектируемого контракта. Не принимает настоящие платежи. Тренировочные данные API не принимает.",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/auth/challenges": {
      "post": {
        "tags": [
          "Вход"
        ],
        "summary": "Запросить код входа или регистрации",
        "operationId": "requestChallenge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Запрос принят; результат следует уточнить",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeOut"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Необходим действующий сеанс",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена или недоступна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния или повторного запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Недопустимые входные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Превышена частота запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/sessions": {
      "post": {
        "tags": [
          "Вход"
        ],
        "summary": "Подтвердить код и открыть сеанс на устройстве",
        "operationId": "openSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Запись создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Необходим действующий сеанс",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена или недоступна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния или повторного запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Недопустимые входные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Превышена частота запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/sessions/current": {
      "delete": {
        "tags": [
          "Вход"
        ],
        "summary": "Завершить текущий сеанс",
        "operationId": "closeSession",
        "responses": {
          "204": {
            "description": "Операция выполнена, тело ответа отсутствует"
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Необходим действующий сеанс",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена или недоступна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния или повторного запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Недопустимые входные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Превышена частота запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/me": {
      "get": {
        "tags": [
          "Вход"
        ],
        "summary": "Получить свою учётную запись",
        "operationId": "getAccount",
        "responses": {
          "200": {
            "description": "Операция выполнена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Необходим действующий сеанс",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена или недоступна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния или повторного запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Недопустимые входные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Превышена частота запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/plans": {
      "get": {
        "tags": [
          "Оплата"
        ],
        "summary": "Получить условные тарифы",
        "operationId": "listPlans",
        "responses": {
          "200": {
            "description": "Операция выполнена",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Plan"
                  },
                  "type": "array",
                  "title": "Ответ операции API"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Необходим действующий сеанс",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена или недоступна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния или повторного запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Недопустимые входные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Превышена частота запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payments": {
      "post": {
        "tags": [
          "Оплата"
        ],
        "summary": "Создать или повторить запрос оплаты",
        "operationId": "createPayment",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 64,
              "title": "Ключ повторяемости запроса"
            },
            "description": "Ключ повторяемости запроса"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Запрос принят; результат следует уточнить",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Некорректный запрос"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Необходим действующий сеанс"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Доступ запрещён"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Запись не найдена или недоступна"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Конфликт состояния или повторного запроса"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Недопустимые входные данные"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Превышена частота запросов"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Сервис временно недоступен"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountOut": {
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid",
            "title": "Идентификатор учётной записи",
            "description": "Идентификатор учётной записи. Идентификатор поля в формате обмена: accountId."
          },
          "email": {
            "type": "string",
            "title": "Адрес электронной почты",
            "description": "Адрес электронной почты. Идентификатор поля в формате обмена: email."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accountId",
          "email"
        ],
        "title": "Учётная запись"
      },
...
      "ChallengeIn": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
            "title": "Адрес электронной почты",
            "description": "Адрес электронной почты. Идентификатор поля в формате обмена: email."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email"
        ],
        "title": "Запрос кода входа"
      },
      "ChallengeOut": {
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid",
            "title": "Идентификатор запроса кода входа",
            "description": "Идентификатор запроса кода входа. Идентификатор поля в формате обмена: challengeId."
          },
          "expiresInSeconds": {
            "type": "integer",
            "title": "Срок действия в секундах",
            "description": "Срок действия в секундах. Идентификатор поля в формате обмена: expiresInSeconds."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "challengeId",
          "expiresInSeconds"
        ],
        "title": "Выданный запрос кода"
      },
...
      "Error": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Код",
            "description": "Код. Идентификатор поля в формате обмена: code."
          },
          "message": {
            "type": "string",
            "title": "Сообщение для пользователя",
            "description": "Сообщение для пользователя. Идентификатор поля в формате обмена: message."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Ошибка операции"
      },
      "LoginIn": {
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid",
            "title": "Идентификатор запроса кода входа",
            "description": "Идентификатор запроса кода входа. Идентификатор поля в формате обмена: challengeId."
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Код",
            "description": "Код. Идентификатор поля в формате обмена: code."
          },
          "deviceLabel": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Название устройства",
            "description": "Название устройства. Идентификатор поля в формате обмена: deviceLabel."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "challengeId",
          "code",
          "deviceLabel"
        ],
        "title": "Подтверждение входа"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Ключ сеанса передаётся в заголовке Authorization: Bearer <ключ>. Значение не публикуется в журнале."
      }
    }
  }
}